Create LongestCommonPrefix
That name sure was loooong, which means it was tiring to type, which means a whole lotta good energy was wasted. Don don don!!! *dramatic affect. screen turns black.* And since we're doing leetcode, which needs you to return something rather than print it, i'll need to put the assertion tests to debugg it aand yeaaaaaaaaah. Hope this finds this informative. Okay byeeeeee (fun fact: school starts tomorrow for me. yes, that was so fasinating). Thank you.